+++ src/main/java/wooteco/subway/dao/section/JdbcSectionDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 @Repository
 public class JdbcSectionDao implements SectionDao {
