+++ src/main/java/wooteco/subway/dao/section/SectionDao.java
@@ -1,7 +1,7 @@
 package wooteco.subway.dao.section;
 
 import java.util.List;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 public interface SectionDao {
