+++ src/main/java/wooteco/subway/dao/station/StationDao.java
@@ -1,7 +1,7 @@
 package wooteco.subway.dao.station;
 
 import java.util.List;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public interface StationDao {
