+++ src/main/java/wooteco/subway/domain/line/Line.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 public class Line {
