+++ src/main/java/wooteco/subway/domain/line/Lines.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 import java.util.List;
 import wooteco.subway.exception.SubwayException;
