+++ src/main/java/wooteco/subway/domain/station/Stations.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.station;
 
 import java.util.List;
 import java.util.stream.Collectors;
