+++ src/main/java/wooteco/subway/dto/station/StationResponse.java
@@ -1,6 +1,6 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.station;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class StationResponse {
