+++ src/test/java/wooteco/subway/acceptance/AcceptanceTest.java
@@ -11,8 +11,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
-import wooteco.subway.dto.LineRequest;
-import wooteco.subway.dto.StationRequest;
+import wooteco.subway.dto.line.LineRequest;
+import wooteco.subway.dto.station.StationRequest;
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
