+++ src/test/java/wooteco/subway/acceptance/SectionAcceptanceTest.java
@@ -9,9 +9,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.LineRequest;
-import wooteco.subway.dto.SectionRequest;
-import wooteco.subway.dto.StationRequest;
+import wooteco.subway.dto.line.LineRequest;
+import wooteco.subway.dto.section.SectionRequest;
+import wooteco.subway.dto.station.StationRequest;
 
 @DisplayName("지하철 노선 관련 기능")
 public class SectionAcceptanceTest extends AcceptanceTest {
