+++ src/test/java/wooteco/subway/dao/LineDaoTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.dao.line.JdbcLineDao;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 @JdbcTest
 class LineDaoTest {
