+++ src/test/java/wooteco/subway/dao/SectionDaoTest.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.dao.section.JdbcSectionDao;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 @JdbcTest
 public class SectionDaoTest {
