+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.dao.station.JdbcStationDao;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 @JdbcTest
 class StationDaoTest {
