+++ src/test/java/wooteco/subway/domain/PathTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.section.Section;
 import wooteco.subway.exception.SubwayException;
 
 public class PathTest {
