+++ src/test/java/wooteco/subway/domain/section/SectionTest.java
@@ -1,9 +1,10 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.section.Section;
 import wooteco.subway.exception.SubwayException;
 
 public class SectionTest {
